Put blog and showcase under the same layout route
